[REF] syllabus_minister: change field type

+++ syllabus_minister/models/unit.py
@@ -10,7 +10,7 @@ class Unit(models.Model):
     _order = 'sequence,id'
 
     name = fields.Char(string='Name')
-    study_hours = fields.Integer(string='Study hours')
+    study_hours = fields.Float(string='Study hours')
     description = fields.Html(string='Description')
     course_id = fields.Many2one('syllabus_minister.course',string='Course')
     sequence = fields.Integer(string='sequence',default=1)
